Modify: 담기,찜하기 라우팅 통합
